Opal Ice Maker Production/Progress Feature (#17)
## :recycle: Current situation

The initial Opal Ice Maker PR included basic support for controlling the
power state and the nightlight state via homekit.

## :bulb: Proposed solution

The changes in this PR introduces a new Ice Production Progress feature.
The Opal Ice Maker exposes an OIM_PRODUCTION value that roughly
translates to the percentage of ice production on a scale from 0 - 100.
As the ice maker produces ice, this value raises sequentially.

With this value I was able to introduce two new features:

1. In homekit, a read only progress bar that updates automatically as
ice is produced.

I struggled a bit with finding the best representation of "progress" and
settled on a Fan Service with the RotationSpeed Characteristic,
admittedly it looks like something the user can set, but its actually
readonly.


![IMG_5449](https://github.com/user-attachments/assets/095caf5b-43a7-4175-9d21-7faa15d52368)


![IMG_5450](https://github.com/user-attachments/assets/c8543c31-daef-4986-aa69-978c0d077746)

2. The ability for the user to set a desired ice production limit in the
plugin settings, such that the ice maker shuts off automatically at the
given limit rather than filling all the way to 100%.

Tying the two features together - the progress bar displays relative to
the users preferred production limit. So for example if the users
preferred production limit is 20, then the progress bar dynamically
increments in steps of 5, such that if the current production value is
at 10, then the progress bar shows 50%.

The default production limit is 100.

In the above photos, I have my production limit set to 50, and for
demonstration purposes hardcoded 30 as the production value. So the
progress bar shows at 60%.

On the plugin side, I had to make updates to how the refreshRate is
handled, to expose the homekit updateRate, and to find a clever way to
pass metadata from the service to the plugin such that the user could
set their desired production limit.

* Update Rate now exposed, minimum value of 30, allowing the user to set
this lower than the default facilitates a more "live" presentation of
the progress bar in the home app. The default seemed to be somewhere
between 2-5 minutes and was painful. Ideally the deviceRefreshRate and
updateRate are the same, so that when homebridge polls the ice machine
that it in turn automatically updates home app.

<img width="621" alt="Screenshot 2025-02-26 at 10 07 26 AM"
src="https://github.com/user-attachments/assets/e0d624b1-05c3-4f24-8e6f-656c07ac33ba"
/>

* I defined a "ProductData" Characteristic on the AccessoryInformation
Service that is passed to the plugin index.html file. Its a bit hacky as
the only way I could figure out how to make it compliant with HomeKit
was to pass the value as a query string, like this
`l=Production_Limit&i=OPL&t=number&d=100`, and parse it in the
javascript into a JSON object. `l` is label, `i` is indicator (or key)
for the option key value pair, `t` is type and `d` is default as in
default value. It seems that there was a constraint on the length of
this value so I had to use letters.

<img width="757" alt="Screenshot 2025-02-26 at 10 10 09 AM"
src="https://github.com/user-attachments/assets/bd0d0234-b9f2-4c58-8cde-62f6093dd547"
/>

## :gear: Release Notes

Opal Ice Machine Production Progress Feature

## :heavy_plus_sign: Additional Information

https://developers.homebridge.io/#/

Because most users will probably be inclined to set their production
limit and then forget it as it runs on a local server persistently, I
did consider giving the user the ability to dynamically _set_ the
progress bar to their desired production limit such that if they dragged
it to 50% full example, that the machine would start and the progress
bar would begin to _decrease_ until it reached 0. My reservation with
this is that it would get complicated if they then wanted to set the
production limit higher or especially lower than they originally
intended. And because the fan rotation speed/progress bar does not
expose details that would educate the user I decided to just keep it
simple.

### Testing

BDD still, I am especially pleased to find that because of the polling
interval, the production limit feature applies not only when the power
is turned on from the apple home app, but also when the user
instantiates ice production by pressing the power button on the machine.
Basically anytime the production threshold is surpassed the machine
shuts off.

Of note, I wasn't completely clear what OIM_PRODUCTION meant but went
with the assumption of a percentage of ice production from 0 - 100. In
reality the production probably gets up to about 120 before the machine
actually shuts off.
Via observation and a stopwatch, I was at least able to rule out that
OIM_PRODUCTION meant number of minutes, because it takes longer than 140
minutes for the machine to fill.


### Reviewer Nudging

Take a look, and see if you agree with the production limit field living
in the device section. It could alternatively live within advanced
options, but in my opinion it should be particular to the device and not
the plugin. Keep in mind that Advanced Option section is only exposed if
one codes for a ProductionData characteristic as part of the
AccessoryInformation service on a given device, its meant to be
extendable basically.

I'd be eager to at least get these changes on a beta release tag for
feedback purposes, but also fine to merge into latest as well.